requirements.txt
pandas
requests
yfinance
SQLAlchemy
psycopg2-binary
python-dotenv
PyYAML
